preserve consumer-provided "runOnJobQueue" value when set to True (#23)
* preserve consumer-provided "runOnJobQueue" value when set to True

Previously, the "runOnJobQueue" parameter was always overwritten with False, even if a consumer explicitly passed True. This behavior prevented consumers from skipping the initial attempt without the job queue, even when it was certain their query would go to the job queue.

By using setdefault(), the parameter is only set to False if it is not already provided. This change allows consumers who know their query requires the job queue to bypass the first attempt and directly push the job to the queue.

* add logic for skipping first query and query directly over job queue